Exercise 8 complete
List component uses map function to render a  ListEntry component for each todo in the todos array in state.  Created functions that allow the user to edit a todo, submit the new todo to the server in a patch request, and rendered the updated todos in the List.jsx component.